Fix tests, minor cleanup
