+++ custom_components/volkswagencarnet/util.py
@@ -1,4 +1,5 @@
-from typing import Optional
+import logging
+from typing import Optional, Union
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
@@ -6,27 +7,32 @@ from homeassistant.helpers import device_registry
 from homeassistant.helpers.device_registry import DeviceRegistry, DeviceEntry
 from volkswagencarnet.vw_vehicle import Vehicle
 
-import const
-from custom_components.volkswagencarnet import DOMAIN
-from error import ServiceError
-from services import _LOGGER
+from .const import CONF_SCANDINAVIAN_MILES, CONF_NO_CONVERSION, DOMAIN
+from .error import ServiceError
+
+_LOGGER = logging.getLogger(__name__)
 
 
 def get_convert_conf(entry: ConfigEntry) -> Optional[str]:
+    """Convert old configuration."""
     return (
-        const.CONF_SCANDINAVIAN_MILES
-        if entry.options.get(const.CONF_SCANDINAVIAN_MILES, entry.data.get(const.CONF_SCANDINAVIAN_MILES, False))
-        else const.CONF_NO_CONVERSION
+        CONF_SCANDINAVIAN_MILES
+        if entry.options.get(CONF_SCANDINAVIAN_MILES, entry.data.get(CONF_SCANDINAVIAN_MILES, False))
+        else CONF_NO_CONVERSION
     )
 
 
-async def get_coordinator(hass: HomeAssistant, device_id: str):
-    """Get the VolkswagenCoordinator."""
+async def get_coordinator_by_device_id(hass: HomeAssistant, device_id: str):
+    """Get the ConfigEntry."""
     registry: DeviceRegistry = device_registry.async_get(hass)
     dev_entry: DeviceEntry = registry.async_get(device_id)
 
-    # Get coordinator handling the device entry
     config_entry = hass.config_entries.async_get_entry(list(dev_entry.config_entries)[0])
+    return await get_coordinator(hass, config_entry)
+
+
+async def get_coordinator(hass: HomeAssistant, config_entry: ConfigEntry):
+    """Get the VolkswagenCoordinator."""
     if config_entry.domain != DOMAIN:
         raise ServiceError("Unknown entity")
     coordinator = config_entry.data.get(
@@ -53,3 +59,24 @@ def get_vehicle(coordinator) -> Vehicle:
     if v is None:
         raise Exception("Vehicle not found")
     return v
+
+
+def validate_charge_max_current(charge_max_current: Optional[Union[int, str]]) -> Optional[int]:
+    """
+    Validate value against known valid ones and return numeric value.
+
+    Maybe there is a way to actually check which values the car supports?
+    """
+    if (
+        charge_max_current is None
+        #  not working # or charge_max_current == "max"
+        or str(charge_max_current) in ["5", "10", "13", "16", "32", "reduced", "max"]
+    ):
+        if charge_max_current is None:
+            return None
+        elif charge_max_current == "max":
+            return 254
+        elif charge_max_current == "reduced":
+            return 252
+        return int(charge_max_current)
+    raise ValueError(f"{charge_max_current} looks to be an invalid value")
